ADD: Added an excel file with the survey breakdown
This excel file has the question tree in the first tab and color-codes
each group of answered questions by decision tree. Each question’s
statistics and breakdowns are in a different tab. There are 55 questions
in total but a third of them have been removed because they contained
no responses.

TODO: the graded responses still need to be reversed.